Implement auto lock when app is idle (#29)
* Introduce PinScreen

- Create a PinScreen widgets
- Setup basic widget structure (i.e. likely child widgets)
- Rely on tbd textTheme's for styling

* Implement PinKeyboard

This commit includes an initial implementation of the PinKeyboard,
key styles are based on a 'to be defined' textstyle.

* Add Pin Bloc

* Implement Basic Bloc Functionality

- Add initial states and events
- Render UI based on these states
- Provide bloc to PinScreen

* Add Basic Error State

* Update Analysis Rules

- Enable prefer single quoted strings
- Update transitive imports
- Resolve double quoted string issues

* Clean up PinKeyboard

* Fix Navigation Crash & Add Dashboard Route

- Fix a crash in the state route state restoration,
which occurred when returning to the app after process
death. Fixed by migrating to onGenerate route methods
to handle basic navigation
- Add a placeholder DashboardRoute to navigate to from
the current PinScreen to avoid getting stuck in validated
state.

* Add 'attemps' and clean up PinScreen

- Add the notion of attempts to show in the PinInvalid state
- Clean up the PinScreen by breaking up the widget tree in
multiple widgets
- Add a mocked attempts usecase

* Add english localization

* Fix bug in attempts reset function

* Clean up imports and re-order methods in pin_screen.dart

* Autolock when app goes into background

* Update l10n rules

- Disable the nullable-getter so we can remove the bang
operator when fetching the translations
- Make `pinScreenAttemptsCount` count type explicit

* Update PinField

- Remove unused error parameter
- Make kDotSize private to class
- Use kDefaultAnimationDuration for dot animation

* Rename `Pin` state & event classes

Use a consistent prefix (Pin) for both event
and state classes and rely on the naming conventions
as defined here https://bloclibrary.dev/#/blocnamingconventions

* Create dedicated usecase directory

Move usecase related classes to a dedicated
/domain/usecase directory, as usecases are likely to
be used throughout multiple features.

* Update PinScreen Design

- Rely on theme's error color
- Update to new margins
- Add arrow in forgot code button

* Update Translations

- Add dutch translations
- Fix issue with '1' case not rendering (wrong syntax)

* Implement PR Feedback

- Rename directory widgets -> widget
- Move Repository to data directory

* Configure default route security

- Default to SecuredRoute for any newly added route
- Add basic wallet_routes_test to verify this behaviour

* Add prefer_relative_imports rule

* Implement delayed background lock

Do not lock the app instantly when going into the background
but instead start a timer that makes sure the app will be
locked after a specified amount of time has passed with
the app in the background.

* Add FLAG_SECURE

Add the FLAG_SECURE flag to the window on android
to avoid screenshots and blank out the app in the
recents menu. Note, blanking out does not take effect
when going into recent menu directly from the wallet
app. But once the user switches to another app or to
the launcher, going into the recents menu again should
show the app in a blanked out state.

* Document auto_lock_observer.dart

* Refactor PinScreen

Instead of having the slightly obscure `navigateOnUnlock`
parameter on the PinScreen this commit changes it to have
an `onUnlock` callback. Giving the responsibility to the
creator of this class to decide what happens when the app
is successfully unlocked.

* Update kBackgroundLockTimeout

Update the kBackgroundLockTimeout to match the requirements

* Perform minor refactor

Rename methods in auto_lock_observer to be more
descriptive.

* Update mock_wallet_repository

- Prefer the seeded() constructor over initializing
the _locked subject in the constructor.
- Update isLockedStream to use the distinct operator,
to avoid unnecessary rebuilds on e.g. failed pin attempts.

* Implement Idle Lock

Introduce an auto lock after 5 minutes of inactivity in the app

* Remove redundant comment