Add basic scanning functionality (#37)
* Introduce QrScreen

Setup an initial QrScreen with tab bar
to switch between code and scan mode and
display placeholder tab widgets.

* Update design system

- Update primaryColor
- Update primaryTextColor

* Fix translations

Was using dutch translations in the english arb file,
this is fixed and dutch translations were added.

* Setup Basic ScanQrTab

- Add mobile_scanner dependency
- Add MobileScanner Widget to ScanQrTab
- Add placholder 'how to' button

* Create and provide ScanQrBloc

* Check in Podfile

Check in the Podfile and other files needed to built for
iOS with the new camera dependency.

* Setup how to scan sheet

Setup the how to scan sheet with some basic
copy. Opted to define the 'drag handle' color
in the theme by means of the divider color for
now. If it turns out we need a dedicated, differnt
divider color we can look at other ways to provide
this color so that it will adapt with the theme.

* Add mock flashlight toggle button

- Add the flashlight toggle (mocked) button

* Update compile and targetSdk

Update the compile and targetSdk to match the latest
version of Android, had to be upgraded for the mobile_scanner
package as it required 32 as a minimum.

* Add flash toggle support

* Clean up FlashToggle

- Remove redundant 'requestedState' from FlashlightToggled
- Move FlashlightStateProxy to dedicated file

* Add initial scanner states

- Generic error state
- Permission needed state
- Perform dummy action (show bottom sheet) on successful scan

* Handle camera permissions

Initially aimed to handle permission solely
using the mobile_scanner package but callbacks
were to limiting for the desired flow. Now added the
permission_handler package to handle permissions, which
should actually make migration from mobile_scanner if desired
an easier process.

All permission flows should now be handled, including automatic
resuming of the camera when the permission was enabled from the
settings page (i.e. when no direct feedback by means of the
Future<bool> request() can be provided). This is done by checking
onResume in this case using the check_permission_on_resume.dart

* Introduce TextArrowButton and add translations

- The TextButton that includes a small arrow is
coming back in multiple screens. As such I opted
to create a TextArrowButton that encapsulates this
visual pattern.
- Add translations for qr permission flow and error state
- Fix lint issues

* Navigate on successful scan event

- Add a VerificationScreen that is used as a dummy destination
- Add success state that is shown after a successful scan,
this includes a CTA to restart the scanner
- Disable 'secure flag' on debug builds to allow screen recording

* Add vibration feedback

Vibrate the device when a QR-code is recognized

* Rename classes and function

Rename some classes and parameters to be more
explicit about their functionality.

* Update 'scan another' UI

* Implement PR Feedback

* Move _kScannerBgColor

Move the color into the ThemeData by means of the
on `secondaryContainer` property.