Add initial verification flow (#43)
* Introduce VerificationRequest

Introduce the notion of a verification request
that can be scanned and pass it along to the
verification screen when navigating.

* Simplify RouteArguments and Setup VerificationBloc

- Setup a QrRequest that models the QR code that triggers a mock flow
- Move loading of the VerificationRequest into a usecase and postpone
this to the VerificationScreen

* Add dynamic back button

Make it so that the back button only shows
when the user is on page > 0.

* Configure desired back and close button behaviour

* Introduce VerificationScreen

* Manage button navigation placement

Migrate to a CustomScrollView so that we can use
the SliverFillRemaining sliver to manage the bottom
placements. This allows us to align the buttons at
the bottom of the screen when the scroll content
does NOT fill the screen, and neatly stack them underneath
the content if the content DOES fill the whole screen.

* Add initial mock data

Add some mock data to built the verification screen
and improve styling to start matching the figma design.

* Improve VerificationRequest Model

Flesh out the model to mock the requested attributes
and some policy information.

* Add ConfirmDataAttributesPage

* Add VerificationSuccessPage

- Add the confirmation page
- Clean up some logic related to showing/hiding
the back button in the app bar.

* Update VerifierPolicy Model

* Update Verification Buttons

The design was updated, this commit aligns it
with the latest design found in Figma.

* Update ConfirmVerifierPage Design

* Update ConfirmDataAttributesPage Design

* Extract Verification Widgets

Extract attribute and policy row into dedicated
widgets.

* Remove Unused TextArrowButton Parameter

* Perform localization

Add initial dutch and english translations.

* Move Mocks to MockVerificationRepository

* Rename VerificationLoadRequest Event

* Hide RippleFeedback on Cards

The cards showed a very tiny ripply feedback when
clicked (in the corners). This commits sets the
border radius to match the cards to fix this minor
visual glitch.

* Forward arguments when constructing bloc

* Revert "Update Verification Buttons"

This reverts commit aeeeb6430925c39a299c4e7befa7f597e4024114.

# Conflicts:
#	lib/src/feature/verification/page/confirm_data_attributes_page.dart
#	lib/src/feature/verification/page/confirm_verifier_page.dart

* Revert "Update ConfirmVerifierPage Design"

* Revert "Update ConfirmDataAttributesPage Design"

* Introduce LinkButton

Add a `LinkButton`, useful when in need of a button
that aligns with other text and doesn't show the
ripple effect.

Also added some missing localizations

* Update LockTimeouts in Debug Builds

* Update Dutch Translations

* Perform refactoring

- Add documentation to link_button.dart
- Remove unused method from confirm_data_attributes_page.dart
- Add documentation to visibility_cubit_back_button.dart
- Extract confirm_buttons.dart into dedicated widget
- Fix some translations

* Simplify LinkButton

No need for the dynamic arrow placement atm.

* Add documentation & consts

- Add some documentation to a method
in the verification_screen.dart to make
it more clear why it's needed.
- Make the size of the success icon a const.

* Add Stepper Indicator

* Implement Initial PR Feedback

* Implement Initial PR Feedback

- Rename ScanQrXXX -> QrScanXXX
- Rename DataAttributeRow
- Minor event/state renames

* Move QR logic

Move the QR decoding logic to the repository layer.
Also slightly expanded the data encapsulated by the QrRequest.
This now indicates what the intended purpose of the Qr code is,
so we will be able to start the correct flow accordingly.
The 'sessionId' is still just a representation of the id hardcoded
in the MockVerificationRepository, but this logic can now be
easily swapped out with a different implementation.

* Make Verifier assets local

Instead of doing a network request to fetch the verifier
logo we embed them in the app for the demo.

* Implement PR Feedback