Fix naming issues (#67)
* Fix incorrect _createXxRoute naming

Rename them to _createXxScreenBuilder as they are
actually returning a WidgetBuilder that is responsible
for building the screen.

* Fix old verifierRequestConditions naming

The string weren't updated when the screen was renamed, and
thus were still prefixed with verifierRequestConditions. This
is now fixed by prefixing them with the verifierPolicy prefix.

* Hide 'Back Button' on PinScreen

When the pin overlay appears on top of e.g.
the verification flow it automatically ads the
back button because the route can be popped. But
it would just pop to another pin screen which is
unwanted. As such disabling this behaviour by proving
a blank widget instead.