Show when requested data is not in Wallet (#65)
* Add state to VerificationScreen

Add some state to the screen to handle new
loading/approved/denied states.

* Style Decline BottomSheet

* Add VerificationDeniedPage

* Introduce ConfirmActionSheet

Break out the ConfirmActionSheet to have a generic
way to ask for a user's input.

* Implement VerificationDeclinedPage Design

- Implement the latest design
- lokalise string resources

* Perform minor verification_screen.dart cleanup

* Make VerificationScreen Stateful

- Remove some complexity that was caused by trying to
keep the screen stateful (e.g. not being able to properly
instantiate a PageController, so passing it around through
the PrimaryScrollController)
- Restore stepper and back button state (which lives in the cubits)
by observing the PageController after the process is restored.
See '_syncCubitsOnAttach()'.

* Make VerificationScreen Stateful

- Remove some complexity that was caused by trying to
keep the screen stateless (e.g. not being able to properly
instantiate a PageController, so passing it around through
the PrimaryScrollController)
- Restore stepper and back button state (which lives in the cubits)
by observing the PageController and storing it's page state in the
_currentPage variable.

* Dispose _currentPage

Make sure the _currentPage RestorableDouble
is disposed properly.

* Update VerificationDeclinedPage

- Add new copy
- Rely on primaryColorDark for StatusIcon

* Update VerificationDeclinedPage

- Add new copy
- Rely on primaryColorDark for StatusIcon
- Update design

* Perform final pre-pr tweaks

* Add MissingAttributesPage to verification flow

- Add the MissingAttributesPage
- Add the missing data mock request
- Add translations

* Add placeholders

Setup a re-usable PlaceholderScreen and
make buttons without any destination navigate
to this screen to demonstrate it's still work
in progress.

* Remove unused VoidCallback

* Update MissingAttributesPage

Remove unused field.

* Update DEMO.md with new QR code