Fix Process Death Crash (#122)
* Ensure all pushed consecutive routes are restorable

* Remove WalletApp RestorationScopeId

Because of the provided restorationScopeId the
MaterialApp attempts to restore the state of the
navigator after process death. This would normally
be the desired behavior but because all data is
currently (for the demo) kept in memory, this results
in the app being in an invalid state (if this happens
after the wallet creation). To circumvent this issue
we now simply disable state restoration by removing
the restorationScopeId from the `MaterialApp` widget.

Once we move to proper persistence for the wallet we should
be able to re-introduce this id to restore proper state
restoration.