Fix 'no cards selected' bug (#140)
When the user doesn't select any cards to be added during
the issuance flow, the app would show the RSOD because it
can't render that state. Now fixed by skipping ahead to
the terminal state, as no other cards have to be added
which is identical to skipping the step.