Fix Duplicate Qr Scans (#178)
When multiple QR codes are in frame, the
issuance/verification/sign flow would be
triggered multiple times causing them to
be displayed in a stacked manner. This
commit makes sure the scanner only processes
QRs if it's ready to do so (i.e. not processing
something else). Fixing the issue described above.