Remove LoadingState when transitioning to StoppedState (#181)
When transitioning to the stopped state in the
sign/verify/issue flow there was an intermediary
loading state. One of the issues here is that transition
to the stopped state could be triggered from this (otherwise
useless) stopped state, causing a loop.. To avoid this
I could either add checks/another flag to make sure
transitioning to stopped from this specific loading state
is no longer possible. But alternatively this specific loading
state could just be skipped. I opted for the latter one, as I
think it keeps the code a lot cleaner, and afaik there was never
an explicit requirement to have a loading state in between the
stop sheet and terminal screen to begin with.