Increase Icons Size (#194)
Increase the icon size of the TextIconButton and the
LinkButton. Was originally set to 12, but after double
checking Figma the bounding box there is actually set
to 16 (if you go up one layer after selecting the icon).
So that's likely why it was (incorrectly) set to 12 in
the first place.