Update Health Mocks (#202)
Create a dedicated attribute type for the health
card insurance expiry date. This fixes the glitch where:

1. User sets up wallet (without health card)
2. User scans first aid verify QR validate
3. It looks like user can provide a health card
expiry date, but this is in fact just an expiry
date from another card.

Desired (and implemented with this fix): When no
health insurance card is added yet, no expiry date
should be provided.